+++ route_verification/bgp/src/tests/cmp.rs
@@ -111,7 +111,7 @@ fn stats() -> Result<()> {
     Ok(())
 }
 
-fn expected_stats() -> [HashMap<usize, AsStats>; 1] {
+fn expected_stats() -> [HashMap<u64, AsStats>; 1] {
     [
         hashmap! {3130=> AsStats { import_ok: 0, export_ok: 0, import_skip: 1, export_skip: 0, import_err: 0, export_err: 0 }, 1239=> AsStats { import_ok: 0, export_ok: 0, import_skip: 1, export_skip: 1, import_err: 0, export_err: 0 }, 9583=> AsStats { import_ok: 0, export_ok: 0, import_skip: 0, export_skip: 0, import_err: 0, export_err: 1 }, 2914=> AsStats { import_ok: 0, export_ok: 0, import_skip: 0, export_skip: 1, import_err: 1, export_err: 0 }},
     ]
